Use the correct enableEncryption() method
