+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/api/session/room/state/StateService.kt
@@ -28,11 +28,6 @@ interface StateService {
      */
     fun updateTopic(topic: String, callback: MatrixCallback<Unit>)
 
-    /**
-     * Enable encryption of the room
-     */
-    fun enableEncryption(algorithm: String, callback: MatrixCallback<Unit>)
-
     fun getStateEvent(eventType: String): Event?
 
     fun getStateEventLive(eventType: String): LiveData<Optional<Event>>
