+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/internal/session/room/state/DefaultStateService.kt
@@ -27,7 +27,6 @@ import im.vector.matrix.android.api.session.events.model.EventType
 import im.vector.matrix.android.api.session.room.state.StateService
 import im.vector.matrix.android.api.util.Optional
 import im.vector.matrix.android.api.util.toOptional
-import im.vector.matrix.android.internal.crypto.MXCRYPTO_ALGORITHM_MEGOLM
 import im.vector.matrix.android.internal.database.mapper.asDomain
 import im.vector.matrix.android.internal.database.model.EventEntity
 import im.vector.matrix.android.internal.database.query.descending
@@ -36,7 +35,6 @@ import im.vector.matrix.android.internal.database.query.where
 import im.vector.matrix.android.internal.task.TaskExecutor
 import im.vector.matrix.android.internal.task.configureWith
 import io.realm.Realm
-import java.security.InvalidParameterException
 
 internal class DefaultStateService @AssistedInject constructor(@Assisted private val roomId: String,
                                                                private val monarchy: Monarchy,
@@ -57,7 +55,7 @@ internal class DefaultStateService @AssistedInject constructor(@Assisted private
 
     override fun getStateEventLive(eventType: String): LiveData<Optional<Event>> {
         val liveData = monarchy.findAllMappedWithChanges(
-                { realm ->  EventEntity.where(realm, roomId, eventType).descending() },
+                { realm -> EventEntity.where(realm, roomId, eventType).descending() },
                 { it.asDomain() }
         )
         return Transformations.map(liveData) { results ->
@@ -77,23 +75,5 @@ internal class DefaultStateService @AssistedInject constructor(@Assisted private
                     this.callback = callback
                 }
                 .executeBy(taskExecutor)
-    }
-
-    override fun enableEncryption(algorithm: String, callback: MatrixCallback<Unit>) {
-        if (algorithm != MXCRYPTO_ALGORITHM_MEGOLM) {
-            callback.onFailure(InvalidParameterException("Only MXCRYPTO_ALGORITHM_MEGOLM algorithm is supported"))
-        } else {
-            val params = SendStateTask.Params(roomId,
-                    EventType.STATE_ROOM_ENCRYPTION,
-                    mapOf(
-                            "algorithm" to algorithm
-                    ))
-
-            sendStateTask
-                    .configureWith(params) {
-                        this.callback = callback
-                    }
-                    .executeBy(taskExecutor)
-        }
     }
 }
